requirements.txt
scikit-learn
numpy
scipy
PyYAML
pytest